requirements.txt
numpy
matplotlib
pandas
scipy
statsmodels
# mhar
# pycalib
python-ternary
composition_stats
jupyter
tqdm
torch
torchvision
# pyswarm
